controlador/requirements.txt
aiohttp==3.9.0
PyJWT==2.8.0
python-dotenv==1.0.0
python-socketio==5.10.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.1
